requirements.txt
wheel
uvicorn       
aioredis
aiosmtplib
anyio
alembic
asgiref
async-timeout
blinker
certifi
cffi
charset-normalizer
click
dnspython
email-validator
fastapi
fastapi-mail
fastapi-utils
greenlet
h11
idna
Jinja2
MarkupSafe
packaging
passlib
python-jose
pdfkit
pycparser
pydantic
fastapi_pagination
PyJWT
pyparsing
qrcode
pdfkit
python-decouple
python-dotenv
python-jose[cryptography]
python-multipart
PyYAML
requests
rfc3986
six
sniffio
sortedcontainers
SQLAlchemy
starlette
typing-extensions
uvicorn
watchgod
websockets
validators
pytest
fastapi-pagination
qrcode
authlib
itsdangerous
aiofiles
Pillow
stripe
pandas
mysqlclient
forex-python
gunicorn
celery
getmac
Pillow
mysql-connector-python